requirements.txt
psycopg2-binary
python-dotenv
httpx
ollama
langchain-ollama
langchain_community
langgraph
docker
sentence-transformers
langchain_huggingface